refactor proxy server
