+++ src/main/java/net/pms/network/proxyserver/ProxyServer.java
@@ -16,7 +16,7 @@
  * along with this program; if not, write to the Free Software
  * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  */
-package net.pms.network;
+package net.pms.network.proxyserver;
 
 import java.io.IOException;
 import java.net.ServerSocket;
@@ -26,7 +26,8 @@ import org.slf4j.LoggerFactory;
 
 public class ProxyServer extends Thread {
 	private static final Logger LOGGER = LoggerFactory.getLogger(ProxyServer.class);
-	private int port;
+
+	private final int port;
 
 	public ProxyServer(int port) {
 		this.port = port;
